pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "erniebot"
version = "0.1.0"
[tool.poetry.dependencies]
beautifulsoup4 = "^4.9"
"discord.py" = "^1.3"
flask = "^1.1"
python = "^3.8"
replit = "^1.0"
[tool.poetry.dev-dependencies]
